+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
 
+## [0.0.2] 2017-11-29
+### Added
+- fixer.io for currency conversation
+### Fixed
+- fix currency panicking (index out of range error)
+- fix GetByAmount order and comparing
+- fix adv cash bug, when AmountInUSD emty if txn currency == USD
+
 ## [0.0.1] 2017-11-28
 Initial